requirements.txt
fastapi
uvicorn
langchain
faiss-cpu
openai
python-dotenv
pydantic
langchain-text-splitters
pypdf
python-docx
webvtt-py
sentence-transformers
numpy